feat(initial): initial version
This is the initial MVP of a piece of software that can connect to a UDM
Pro and get the existing hosts that are connected to the system and
create a configurable hosts.txt from that. This file can then be brought
into dnsmasq or CoreDNS to serve domain names for the local domain.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>