fix(parser): Don't fail on invalid IP
In a situation when a UDM Pro has recently rebooted and it knows of a
device from before the reboot but it hasn't seen it since the reboot, it
might return something saying there is no IP address. This would
previously cause the application to crash. This fixes that problem by
changing the logger.Fatalf call to logger.Warnf.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>