feat(persistance): persist IP addresess of disconnected devices
One of the things that was most annoying to me about this system is that
when a device disconnected and no longer showed up in the Unifi API, it
wouldn't show up in the hostnames anymore. This led to a couple of
things - first, it meant that you'd get hostnames not found frequently
when you know the device exists, it just went to sleep. Second, you'll
have to wait until the next polling round to get the new hosts - even if
they normally are given the same IP address.

This fixes that by adding an element of persistence by storing a
lastseen and lastseenUnifi time (although the latter isn't really used).
If a hostname has different IP addresses, it will use the most recent IP
address. If two devices show up with the same IP address, it will use
the one that was seen the most recently. This will flush those entries,
but it's not that big of a deal.

Right now the entries are persisted until the application restarts. In
the future, this should be made so it can bootstrap from a file or the
entries time out after some sufficiently long time.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>