fix(error): Make error handling less aggressive
If the script happens to hit the UDM Pro while it is updating, then it
would crash out. These changes should make it so it error handling is
less aggressive and it just loops around until it can try again.

This isn't that needed if you're running the program so it always
restarts, but if you're caching offline hosts, then you'll lose your
cache in these situations.

Which brings up the bigger issue - the script might want to cache all of
the hosts that it has found in something persistent so it can be brought
back up after a crash.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>