fix(hostnames): filter out hostnames with colons
The recent upgrade on my UDM-Pro resulted in it returning some hostnames
with colons in them. This is bad as these are not valid hostnames (it
was just returning the MAC addresses). This makes it so the program
removes MAC address hostnames by default. Eventually, they might come
around to something more useful.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>