chore(logging): clean up logging
This makes sure that the levels for logging are appropriate - before
there was a Warn that should've been an Info. It also improves the
consistency of the messages, adds a loop counter so you have an idea of
how many times it has run, and provides information about the number of
records it has removed because they were filtered out.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>