feat(sql): optionally write output to an SQL database
As I've started to use CoreDNS with the pdsql plugin to read my data
from an sqlite database (this actually a lot more flexible overall),
this tool should be able to write to a file. This brings in the
initial support for writing to a database and also brings in some long
missing tests to support it.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>