feat(cname): support cname values
This introduces a new directive `cnames` in config.toml that takes a
list of `{cname="BLAH", hostname="BLAH2"}` to create CNAME entries. If
you're using the hostfile, it will do it's best for it and add them
onto the host if it can fully resolve it. With the database, it adds
them as CNAME entries into the database so they can be read by
something like coredns-pdsql.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>