feat(scraper): add support for mac addresses
The application previously had support for mac address filtering, but
if you left them in then it would just include hostnames with `:`
characters in them. We don't want that. This makes it so if
`Processing.keep_macs` in the config is true then the macs will be
kept as hostnames, but the `:` will be replaced by `-`.

Closes #2

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>