feat(hostname): allow additional hosts to override others
Previously, you could have double entires when you defined a host in
the processing.additional block. Now, this adds a new parameter called
`KeepMultiple` which can be set for each host. It defaults to false,
which means that if the name appears in additional hosts, it will not
appear from anything that is pulled up from the UDM. If it is true, it
will keep all entries, potentially resulting in multiple hosts with
the same IP address on your network. But, hey, maybe you wanted that.

Closes #1

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>